角色Locomotion模块重构，与LevelMaster剥离，优化API命名
