+++ jyx2/Assets/Prefabs/Player.prefab
@@ -99,6 +99,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: da4c7218c4d328544ba9f8c445f3fd11, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  _locomotionController:
+    navPointerPrefab: {fileID: 1613908073751994, guid: b14eaf554645b7e4e99d40e234e50930,
+      type: 3}
+    unlockDegree: 10
   m_Animancer: {fileID: 6823037285151171485}
   m_Animator: {fileID: 4683237163795035701}
   IsOnBoat: 0
