+++ jyx2/Assets/Scripts/PlayerController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dcb74ce89604be94b9374c7102e36e4a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
