+++ jyx2/Assets/Scripts/PlayerController/PlayerLocomotionController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 12eede25b186fba45b0dd207ec9362ee
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
